Replaced credentialGenerator with a more broad authenticationGenerator which will handle other types of authentication
